vertex_pipelines/spotify-trigger/requirements.txt
google-cloud-bigquery==3.15.0
google-cloud-aiplatform[pipelines]==1.43.0
